iscolided, and fix

+++ TopGun/IModel.cs
@@ -16,9 +16,10 @@ namespace TopGun
     
     public class OutputCoordinate
     {
+        
         public Coordinate CoordinatePlayer { get; set; }
         public Dictionary <int, Coordinate> CoordinateEnemy { get; set; }
-        public List <(Coordinate pos, int , int )> CoordinateBullet { get; set; }
+        public List <(Coordinate pos, double , int )> CoordinateBullet { get; set; }
         public int Radius { get; set; }
     }
 }
